requirements.txt
requests
pymysql
flaskapp
flask
Flask
gunicorn
twilio
pyyaml